Submission/Submission_3_Kelola_Data/app/src/main/java/com/onedev/dicoding/architecturecomponent/ui/fragment/movie/MovieViewModel.kt
package com.onedev.dicoding.architecturecomponent.ui.fragment.movie

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.paging.PagedList
import com.onedev.dicoding.architecturecomponent.data.source.MovieRepository
import com.onedev.dicoding.architecturecomponent.data.source.local.entity.MovieEntity
import com.onedev.dicoding.architecturecomponent.vo.Resource

class MovieViewModel(private val movieRepository: MovieRepository) : ViewModel() {

    fun getPopularMovie(): LiveData<Resource<PagedList<MovieEntity>>> = movieRepository.getPopularMovie()

}